scrubbed files
Added a profile that points to scrubbed schema and put that as part of config settings in each sql. also split this into two models where one is to form chronometer_scrubbed, and the other makes an aging AR table based on that data.

This is not how DBT is supposed to be used since AR table is defined from the scrubbed table, which is also a dbt created table too. But doing it this way since we want to pretend that the chronometer_scrubbed tables are the source tables (which is why I added them to the models.yml).

Because of that, we will want to schedule the chronometer_scrubbed model to refresh on an interval sooner than aging_ar. But we could atleast schedule chronometer_scrubbed to refresh anytime the maria_rds tables are updated.